cubes/swebench-verified-cube/pyproject.toml
[project]
name = "swebench-verified-cube"
version = "0.1.0"
description = "SWE-bench Verified — 500 real-world GitHub issues with test-based validation"
requires-python = ">=3.12"
dependencies = [
    "cube-standard",
    "datasets",
]

[project.entry-points."cube.benchmarks"]
swebench-verified-cube = "swebench_verified_cube.benchmark:SWEBenchVerifiedBenchmark"

[build-system]
requires = ["uv_build>=0.8,<0.9"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0",
]

[tool.ruff]
fix = true
line-length = 120
indent-width = 4
